Remove all `@inbounds` – this is not a hotspot, nor sufficiently safe
When you're doing this much I/O, bounds checking is a completely trivial
overhead, so having `@inbounds` annotations everywhere is an unnecessary
optimization. There are also lots of instances in this code using indexes
that are by no means certain to be in bounds. I just encountered one that
was causing very hard to debug problems as it was basically trashing the
surrounding memory.