Allow a fix for #249
From some research discovered
[here](http://datamining.togaware.com/survivor/Database_Connection.html)
and in the Netezza ODBC documentation
[here](https://www.ibm.com/support/knowledgecenter/SSULQD_7.1.0/com.ibm.nz.datacon.doc/r_datacon_odbc_driver_setup_win.html),
the Netezza ODBC driver, by default, has a 'Prefetch Count' setting,
defaulting to 256, that limits the # of rows that will return for a
resultset. It's unclear whether you can set this number to -1 or some
really large integer # (4 trillion or so) and if that would work (or try
to consume that much memory by pre-allocating?).

In any case, RODBC has a setting `believeNRows`, so we introduce a
similar (though IMO better named) keyword argument
`ignore_driver_row_count`, which, when set to `true`, will force the
result handling code to iterate rows with no known length, which is the
setting we want when we're unable to pre-allocate full columns.