Finalize statements on connection before closing
Fixes #292. This was a bit of a bear to track down, mainly
because the docs for SQLDisconnect state explicitly that all statements
on the connection will be closed/freed when disconnecting. For whatever reason, that doesn't seem to be the case on windows, because when we keep a `WeakKeyDict` of statement handles and finalize them all before closing the connection, we no longer have issues. ¯\_(ツ)_/¯